// apps/inventario/src/modules/proveedores/enums/proveedor.enums.ts
export enum TipoProveedor {
  PRODUCTOS = 'productos',
  SERVICIOS = 'servicios',
  EQUIPOS = 'equipos',
  MATERIALES = 'materiales',
  LOGISTICA = 'logistica',
  TECNOLOGIA = 'tecnologia',
  OTROS = 'otros',
}

export enum EstadoProveedor {
  ACTIVO = 'activo',
  INACTIVO = 'inactivo',
  SUSPENDIDO = 'suspendido',
  EN_REVISION = 'en_revision',
}